requirements.txt
gunicorn
flask
supabase
python-dotenv